requirements.txt
blinker==1.4
click==6.7
Flask==0.12.2
Flask-DebugToolbar==0.10.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pigpio==1.38
pyserial==3.4
Werkzeug==0.14.1
